feature: delete icon sheet widget, edit plus_button to use OverlayEntry (#25)
